OMG.Utxo.Position <- OMG.InputPointer
