+++ apps/omg/test/omg/state/persistence_test.exs
@@ -127,7 +127,7 @@ defmodule OMG.State.PersistenceTest do
     assert {:ok, tx} == Transaction.Recovered.recover_from(block_tx)
 
     assert {:ok, 1000} ==
-             tx |> Transaction.get_inputs() |> hd() |> InputPointer.to_db_key() |> OMG.DB.spent_blknum()
+             tx |> Transaction.get_inputs() |> hd() |> Utxo.Position.to_input_db_key() |> OMG.DB.spent_blknum()
   end
 
   @tag fixtures: [:alice]
