+++ apps/omg/test/omg/state/utxo_set_test.exs
@@ -41,7 +41,7 @@ defmodule OMG.State.UtxoSetTest do
       inputs
       |> Enum.zip(outputs)
       |> Enum.map(fn {input, output} -> {input, %Utxo{output: output, creating_txhash: <<1>>}} end)
-      |> Enum.map(fn {input, utxo} -> {InputPointer.to_db_key(input), Utxo.to_db_value(utxo)} end)
+      |> Enum.map(fn {input, utxo} -> {Utxo.Position.to_input_db_key(input), Utxo.to_db_value(utxo)} end)
 
     utxo_set = UtxoSet.init(db_query_result)
