+++ apps/omg_watcher/test/omg_watcher/utxo_exit/core_test.exs
@@ -36,7 +36,7 @@ defmodule OMG.Watcher.UtxoExit.CoreTest do
       encode_utxo = position |> Utxo.Position.encode()
 
       [output] = Transaction.get_outputs(Transaction.Payment.new([], [{alice.addr, @eth, 10}]))
-      fake_utxo_db_kv = {OMG.InputPointer.to_db_key(position), Utxo.to_db_value(%Utxo{output: output})}
+      fake_utxo_db_kv = {OMG.Utxo.Position.to_input_db_key(position), Utxo.to_db_value(%Utxo{output: output})}
 
       assert {:ok,
               %{
